wco_phase1/requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python-headless
google-api-python-client
google-auth
python-dotenv
google-auth-oauthlib